move roll function to utils for test
